+++ types/index.d.ts
@@ -1,0 +1,11 @@
+export declare function useCloud(): string;
+import { Interface } from './Interface.js';
+import { Node } from './Node.js';
+import { OAuth } from './OAuth.js';
+declare const _default: {
+    Interface: typeof Interface;
+    Node: typeof Node;
+    OAuth: typeof OAuth;
+    logger: import("winston").Logger;
+};
+export default _default;