+++ types/logger.d.ts
@@ -1,0 +1,5 @@
+import winston from 'winston';
+declare const poly: winston.Logger;
+export declare const ns: winston.Logger;
+export declare function errorStack(err: any, ...args: any[]): void;
+export default poly;